requirements.txt
### requirements.txt
flask
transformers
sentence-transformers
faiss-cpu
torch